+++ apps/connect/source/ftrack_connect/utils/plugin.py
@@ -161,14 +161,16 @@ def fetch_github_releases(latest=True, prereleases=False):
     version of each plugin is returned. If *prereleases* is True,
     prereleases are included in the result.'''
 
-    logger.debug(f'Fetching releases from: {INTEGRATIONS_REPO}')
+    logger.debug(
+        f'Fetching releases from: {INTEGRATIONS_REPO} (pre-releases: {prereleases})'
+    )
 
     response = requests.get(f"{INTEGRATIONS_REPO}/releases")
     if response.status_code != 200:
         logger.error(f'Failed to fetch releases from {INTEGRATIONS_REPO}')
         return []
 
-    data = []
+    data = {}
 
     # Expect list of releases
     for release in response.json():
@@ -194,7 +196,10 @@ def fetch_github_releases(latest=True, prereleases=False):
             continue
 
         if not prereleases and release.get('prerelease') is True:
-            logger.debug(f'   Skipping prerelease: {tag_name}')
+            logger.debug(f'   Skipping pre-release: {tag_name}')
+            continue
+        elif prereleases and not release.get('prerelease'):
+            logger.debug(f'   Skipping release: {tag_name}')
             continue
         release_data = {
             'id': release['id'],
@@ -241,23 +246,17 @@ def fetch_github_releases(latest=True, prereleases=False):
         if url:
             logger.debug(f'Supplying release: {tag_name}')
             release_data['url'] = url
-            data.append(release_data)
+            if not latest or package not in data:
+                data[package] = release_data
+            else:
+                current_version = data[package]['tag'].rsplit('/', 1)[1][1:]
+                if parse(current_version) < parse(version):
+                    # This version is higher
+                    data[package] = release_data
 
-    if latest:
-        # Only provide the latest version
-
-        data.sort(key=lambda x: x['tag'], reverse=True)
-
-        result = []
-        for item in data:
-            if (
-                not result
-                or item['tag'].rsplit('/', 1)[0]
-                != result[-1]['tag'].rsplit('/', 1)[0]
-            ):
-                result.append(item)
-    else:
-        result = data
+    result = []
+    for release in list(data.values()):
+        result.append(release)
 
     return result