+++ libs/qt/source/ftrack_qt/widgets/icons/status_icon.py
@@ -3,7 +3,10 @@
 # TODO: clean this code
 import logging
 
-from Qt import QtCore, QtWidgets, QtGui, QtSvg
+try:
+    from PySide6 import QtWidgets, QtCore
+except ImportError:
+    from PySide2 import QtWidgets, QtCore
 
 import ftrack_constants as constants
 from ftrack_qt.widgets.icons import MaterialIcon