+++ libs/qt/source/ftrack_qt/widgets/thumbnails/asset_version_thumbnail.py
@@ -1,7 +1,10 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2024 ftrack
 
-from Qt import QtCore, QtGui, QtWidgets
+try:
+    from PySide6 import QtCore, QtWidgets
+except ImportError:
+    from PySide2 import QtCore, QtWidgets
 
 from ftrack_qt.widgets.thumbnails.base_thumbnail import ThumbnailBase
 
@@ -19,8 +22,8 @@ class AssetVersionThumbnail(ThumbnailBase):
         if self._scale:
             scaled_pixmap = pixmap.scaled(
                 self.size(),
-                QtCore.Qt.KeepAspectRatio,
-                QtCore.Qt.SmoothTransformation,
+                QtCore.Qt.AspectRatioMode.KeepAspectRatio,
+                QtCore.Qt.TransformationMode.SmoothTransformation,
             )
         else:
             scaled_pixmap = pixmap