+++ libs/qt/source/ftrack_qt/widgets/user/ftrack_user.py
@@ -1,7 +1,10 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2024 ftrack
 # TODO: Clean this code
-from Qt import QtCore, QtWidgets, QtGui
+try:
+    from PySide6 import QtWidgets, QtCore
+except ImportError:
+    from PySide2 import QtWidgets, QtCore
 
 from ftrack_qt.widgets.thumbnails import UserThumbnail as UserThumbnail
 
@@ -18,15 +21,13 @@ class FtrackUser(QtWidgets.QFrame):
         super(FtrackUser, self).__init__(parent=parent)
         self.session = session
 
-        self.setObjectName('ftrack-userid-widget')
-
         self.pre_build()
         self.build()
 
     def pre_build(self):
         self.setLayout(QtWidgets.QHBoxLayout())
         self.layout().setContentsMargins(0, 0, 0, 0)
-        self.layout().setAlignment(QtCore.Qt.AlignRight)
+        self.layout().setAlignment(QtCore.Qt.AlignmentFlag.AlignRight)
 
     def build(self):
         username = self.session.api_user