+++ projects/framework-photoshop/extensions/widgets/photoshop_image_options_selector.py
@@ -3,7 +3,10 @@
 
 from functools import partial
 
-from Qt import QtWidgets, QtCore, QtGui
+try:
+    from PySide6 import QtWidgets, QtCore
+except ImportError:
+    from PySide2 import QtWidgets, QtCore
 
 from ftrack_framework_qt.widgets import BaseWidget