+++ tools/cookiecutter-framework-project/{{cookiecutter.project_name}}/connect-plugin/hook/discover_ftrack_framework_{{cookiecutter.integration_name}}.py
@@ -1,12 +1,12 @@
 # :coding: utf-8
-# :copyright: Copyright (c) 2014-2024 ftrack
+# :copyright: Copyright (c) 2024 ftrack
 
 import os
 import ftrack_api
 import logging
 import functools
 
-from ftrack_connect.util import get_connect_plugin_version
+from ftrack_utils.version import get_connect_plugin_version
 
 # The name of the integration, should match name in launcher.
 NAME = 'framework-{{ cookiecutter.integration_name }}'