+++ libs/framework-qt/release_notes.md
@@ -1,7 +1,14 @@
 # ftrack Framework QT library release Notes
 
-## v2.1.0rc1
-2024-03-15
+## Upcoming
+
+* [new] PySide6 support.
+* [new] PySide2 support.
+* [changed] Removed Qt.py dependency.
+* [changed] BaseDialog contain stackedWidget inside with a pre-defined overlay widget. Methods to manage stackedWidgets added. 
+
+## v2.1.0
+2024-04-02
 
 * [changed] BaseContextDialog inherit from BaseDialog. 
 * [changed] Removed pyside extra from ftrack-qt dependency.