+++ libs/qt-style/release_notes.md
@@ -1,7 +1,13 @@
 # ftrack Framework QT Style library release Notes
 
-## v2.1.0rc1
-2024-03-15
+## upcoming
+
+* [new] PySide6 support, through build script.
+* [new] PySide2 support, through build script.
+* [changed] Clean up style files to make it easy to maintain. Using properties for overrides by default.
+
+## v2.1.0
+2024-04-02
 
 * [changed] Improve selected items.