+++ libs/utils/release_notes.md
@@ -1,8 +1,11 @@
 # ftrack Utils library release Notes
 
-## v2.0.2rc1
-2024-03-15
+## v2.1.0
+2024-04-02
 
+* [changed] Extensions/environment; Changed the name of FTRACK_EXTENSIONS_PATH environment variable back to FTRACK_FRAMEWORK_EXTENSIONS_PATH.
+* [changed] Extensions/registry; Extension type filter support to get_extensions_from_directory.
+* [new] Get connect plugin version added on Utils
 * [fix] fix track_usage decorator bug. Arguments are correctly queried.