+++ projects/framework-maya/release_notes.md
@@ -1,6 +1,12 @@
 # ftrack Framework Maya integration release Notes
 
 ## Upcoming
-YYYY-mm-dd
 
+* [changed] Replace Qt.py imports to PySide2 and PySide6 on widgets.
+
+## v24.4.0
+2024-04-02
+
+* [changed] maya_export_options_selector widget with all options set to True by default
+* [changed] Upgrade ftrack-framework-core, ftrack-utils and ftrack-qt version.
 * [new] Initial release.